Adapt the path validation to 25.8:
- `resolvePathInsideTable` lives in the flat `DataLakes/Common.{h,cpp}` (25.8
  predates the `Common/` subdirectory layout).
- `TableSnapshot`: apply the validation at 25.8's construction site
  (`context->data_prefix`), and keep `data_files_cv.notify_all()` in the
  `scanCallback` catch (without it the stored scan exception never wakes the
  consumer in `next` and the query hangs; 25.8's void-returning callback cannot
  use master's `return false`).
- `DeltaLakeMetadata`: keep 25.8's path handling (no `Poco::URI::decode`,
  `read_path_string` / `getPathForRead().path` bases), wrap with
  `resolvePathInsideTable`.
- `Iceberg/Utils.cpp`: keep 25.8's metadata-selection code (the master-side
  cache machinery is unrelated to this PR). For the explicit
  `iceberg_metadata_file_path`: relative paths go through
  `resolvePathInsideTable`; the 25.8-only "already table-prefixed" branch now
  lexically normalizes and validates containment, since it previously checked
  only the first path component and embedded `..` could escape the table
  directory.

+++ src/Common/hasLinuxCapability.cpp
@@ -8,7 +8,6 @@
 #include <cstdint>
 #include <base/types.h>
 #include <Common/Exception.h>
-#include <Common/ErrnoException.h>
 
 
 namespace DB
@@ -61,7 +60,7 @@ static Capabilities getCapabilities()
 bool hasLinuxCapability(int cap)
 {
     static Capabilities capabilities = getCapabilities();
-    return (UInt64(1) << cap) & capabilities.effective;
+    return (1 << cap) & capabilities.effective;
 }
 
 }
