+++ src/Processors/Merges/IMergingTransform.h
@@ -189,7 +189,7 @@ protected:
         {
             LOG_DEBUG(log, "{}, {} blocks, {} rows, {} bytes in {} sec., {} rows/sec., {}/sec.",
                 transform_message, stats.blocks, stats.rows, stats.bytes,
-                seconds, static_cast<double>(stats.rows) / seconds, ReadableSize(static_cast<double>(stats.bytes) / seconds));
+                seconds, stats.rows / seconds, ReadableSize(stats.bytes / seconds));
         }
     }
